+++ tests/files/src/0721_false_positive_coalesce.php
@@ -1,0 +1,8 @@
+<?php
+
+function test() {
+    static $a = null;
+    var_export($a ?? ($a = rand(0,10)));  // should not warn about static variables
+    $b = null;
+    var_export($b ?? ($b = rand(0,10)));
+}